+++ packages/flutter/lib/src/material/tooltip.dart
@@ -399,9 +399,9 @@ class _TooltipState extends State<Tooltip> with SingleTickerProviderStateMixin {
 
     // Only check for hovering if there is a mouse connected.
     if (_mouseIsConnected) {
-      result = Listener(
-        onPointerEnter: (PointerEnterEvent event) => _showTooltip(),
-        onPointerExit: (PointerExitEvent event) => _hideTooltip(),
+      result = MouseRegion(
+        onEnter: (PointerEnterEvent event) => _showTooltip(),
+        onExit: (PointerExitEvent event) => _hideTooltip(),
         child: result,
       );
     }
